+++ src/web-index.go
@@ -18,7 +18,7 @@ func index(w http.ResponseWriter, r *http.Request) {
 	guiData.Config = AppConfig
 	guiData.Links = AllLinks
 
-	tmpl, _ := template.ParseFiles("templates/index.html", "templates/header.html", "templates/footer.html")
+	tmpl, _ := template.ParseFS(TemplHTML, "templates/index.html", "templates/header.html", "templates/footer.html")
 	tmpl.ExecuteTemplate(w, "index", guiData)
 }